Some further renames from "intermediate model" to "streamed value".
