+++ platforms/ide/tooling-api/src/main/java/org/gradle/tooling/internal/consumer/connection/InternalBuildActionAdapter.java
@@ -25,7 +25,7 @@ import org.gradle.tooling.internal.consumer.versioning.VersionDetails;
 import org.gradle.tooling.internal.protocol.InternalActionAwareBuildController;
 import org.gradle.tooling.internal.protocol.InternalBuildActionVersion2;
 import org.gradle.tooling.internal.protocol.InternalBuildControllerVersion2;
-import org.gradle.tooling.internal.protocol.InternalObjectRelay;
+import org.gradle.tooling.internal.protocol.InternalStreamedValueRelay;
 
 import java.io.File;
 
@@ -67,7 +67,7 @@ public class InternalBuildActionAdapter<T> implements org.gradle.tooling.interna
 
     private BuildController wrapBuildController(final InternalBuildControllerVersion2 buildController) {
         ProtocolToModelAdapter protocolToModelAdapter = new ProtocolToModelAdapter(new ConsumerTargetTypeProvider());
-        if (buildController instanceof InternalObjectRelay) {
+        if (buildController instanceof InternalStreamedValueRelay) {
             return new StreamingAwareBuildControllerAdapter(buildController, protocolToModelAdapter, new ModelMapping(), versionDetails, rootDir);
         } else if (buildController instanceof InternalActionAwareBuildController) {
             return new NestedActionAwareBuildControllerAdapter(buildController, protocolToModelAdapter, new ModelMapping(), versionDetails, rootDir);
