+++ platforms/ide/tooling-api/src/main/java/org/gradle/tooling/internal/consumer/connection/StreamingAwareBuildControllerAdapter.java
@@ -20,16 +20,16 @@ import org.gradle.tooling.internal.adapter.ProtocolToModelAdapter;
 import org.gradle.tooling.internal.consumer.versioning.ModelMapping;
 import org.gradle.tooling.internal.consumer.versioning.VersionDetails;
 import org.gradle.tooling.internal.protocol.InternalBuildControllerVersion2;
-import org.gradle.tooling.internal.protocol.InternalObjectRelay;
+import org.gradle.tooling.internal.protocol.InternalStreamedValueRelay;
 
 import java.io.File;
 
 public class StreamingAwareBuildControllerAdapter extends NestedActionAwareBuildControllerAdapter {
-    private final InternalObjectRelay relay;
+    private final InternalStreamedValueRelay relay;
 
     public StreamingAwareBuildControllerAdapter(InternalBuildControllerVersion2 buildController, ProtocolToModelAdapter adapter, ModelMapping modelMapping, VersionDetails gradleVersion, File rootDir) {
         super(buildController, adapter, modelMapping, gradleVersion, rootDir);
-        this.relay = (InternalObjectRelay) buildController;
+        this.relay = (InternalStreamedValueRelay) buildController;
     }
 
     @Override
