+++ platforms/ide/tooling-api/src/main/java/org/gradle/tooling/internal/consumer/parameters/ConsumerOperationParameters.java
@@ -211,7 +211,7 @@ public class ConsumerOperationParameters implements BuildParameters {
             this.cancellationToken = cancellationToken;
         }
 
-        public void setIntermediateModelListener(StreamedValueListener streamedValueListener) {
+        public void setStreamedValueListener(StreamedValueListener streamedValueListener) {
             this.streamedValueListener = streamedValueListener;
         }
